refactor(ssr): move `LightningElement` to own file
