+++ packages/@lwc/ssr-runtime/src/class-list.ts
@@ -4,7 +4,8 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
-import { LightningElement } from '.';
+
+import { LightningElement } from './lightning-element';
 
 const MULTI_SPACE = /\s+/g;
