+++ packages/@lwc/ssr-runtime/src/mutation-tracker.ts
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
 
-import { LightningElement } from '.';
+import { LightningElement } from './lightning-element';
 
 export class MutationTracker {
     mutationMap = new WeakMap<LightningElement, Set<string>>();
@@ -26,3 +26,5 @@ export class MutationTracker {
         }
     }
 }
+
+export const mutationTracker = new MutationTracker();
